+++ packages/browser-wallet/src/popup/pages/SignMessage/i18n/en.ts
@@ -1,5 +1,5 @@
 const t = {
-    description: 'Sign message',
+    description: 'Sign message with account {{address}}',
     sign: 'Sign',
     deny: 'Deny',
     error: 'Error',