+++ packages/browser-wallet-api-helpers/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2.0
+
+-   Updated API of sendTransaction and signMessage to require the account address.
+
 ## 0.1.1
 
 -   sendTransaction can now take a 5th argument, which is the schema's version. This will allow V1 contract parameters to be serialized.